Moved/split time-settings and fixed user-settings
